fix model and JPA repositories

+++ sistemaSolar/src/main/java/com/mercadolibre/solarsystem/repository/PlanetRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlanetRepository extends CrudRepository <Planet, Long> {
+public interface PlanetRepository extends CrudRepository <Planet, Integer> {
 
 }
